refactor: component converted to typescript
